Restructure of file paths
Restructured the directory according to the directions of the project lead. 
Change sys.path of MakeFile and Sphinx according to the new structure of the file tree.
Changed program documentation of Log(a,b) to Ln(a). 
Changed tests from Log(a,b) to Ln(a).
Discovered and fixed wrong output values in some tests.